+++ sdk/src/main/java/com/mapbox/maps/MapController.kt
@@ -207,9 +207,6 @@ internal class MapController : MapPluginProviderDelegate, MapControllable {
   }
 
   override fun setMaximumFps(fps: Int) {
-    if (fps <= 0) {
-      return
-    }
     renderer.setMaximumFps(fps)
   }
 
@@ -233,6 +230,18 @@ internal class MapController : MapPluginProviderDelegate, MapControllable {
     renderer.renderThread.renderHandlerThread.post {
       renderer.renderThread.eglCore.removeRendererStateListener(rendererSetupErrorListener)
     }
+  }
+
+  internal fun setScreenRefreshRate(refreshRate: Int) {
+    if (refreshRate <= 0) {
+      logE(
+        TAG,
+        "Screen refresh rate could not be <= 0! " +
+          "Setting max fps and fps counter will not work properly."
+      )
+      return
+    }
+    renderer.renderThread.setScreenRefreshRate(refreshRate)
   }
 
   //