+++ sdk/src/main/java/com/mapbox/maps/MapSurface.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import android.graphics.Bitmap
 import android.view.MotionEvent
 import android.view.Surface
+import android.view.WindowManager
 import com.mapbox.maps.plugin.MapPlugin
 import com.mapbox.maps.plugin.delegates.MapPluginProviderDelegate
 import com.mapbox.maps.renderer.MapboxSurfaceRenderer
@@ -28,7 +29,8 @@ import com.mapbox.maps.renderer.widget.Widget
  * @param mapInitOptions the init options to for map
  */
 class MapSurface @JvmOverloads constructor(
-  context: Context,
+  // could use strong ref here as MapInitOptions have strong ref in any case
+  private val context: Context,
   val surface: Surface,
   mapInitOptions: MapInitOptions = MapInitOptions(context)
 ) : MapPluginProviderDelegate, MapControllable {
@@ -49,6 +51,10 @@ class MapSurface @JvmOverloads constructor(
    */
   fun surfaceCreated() {
     renderer.surfaceCreated()
+    // display should not be null at this point but to be sure we will fallback to DEFAULT_FPS
+    val screenRefreshRate = (context.getSystemService(Context.WINDOW_SERVICE) as WindowManager)
+      .defaultDisplay?.refreshRate?.toInt() ?: MapView.DEFAULT_FPS
+    mapController.setScreenRefreshRate(screenRefreshRate)
   }
 
   /**
@@ -139,9 +145,6 @@ class MapSurface @JvmOverloads constructor(
    * @param fps The maximum fps
    */
   override fun setMaximumFps(fps: Int) {
-    if (fps <= 0) {
-      return
-    }
     renderer.setMaximumFps(fps)
   }